Barcodescanner functional
