+++ RIPD/AppShell.xaml.cs
@@ -12,6 +12,7 @@ namespace RIPD
       #region Routes
       Routing.RegisterRoute(nameof(FoodDetailsPage), typeof(FoodDetailsPage));
       Routing.RegisterRoute(nameof(NewFoodPage), typeof(NewFoodPage));
+      Routing.RegisterRoute(nameof(BarcodeScannerPage), typeof(BarcodeScannerPage));
       #endregion Routes
     }
   }
