+++ RIPD/Pages/NewFoodPage.xaml
@@ -3,6 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:viewmodels="clr-namespace:RIPD.ViewModels"
              xmlns:models="clr-namespace:RIPD.Models"
+             xmlns:views="clr-namespace:RIPD.Views"
              x:Class="RIPD.Pages.NewFoodPage"
              x:DataType="viewmodels:NewFoodVM">
   <Grid RowDefinitions="*,auto">
@@ -12,7 +13,7 @@
       <Entry Placeholder="Manufacturer" Text="{Binding Manufacturer}"/>
       <Entry Placeholder="Description" Text="{Binding Description}"/>
       <Entry Placeholder="Image" Text="{Binding Image}"/>
-    </VerticalStackLayout>
+        </VerticalStackLayout>
     <Button Grid.Row="1" Text="Create Food" Command="{Binding CreateNewFoodCommand}"/>
   </Grid>
 </ContentPage>
