+++ RIPD/ViewModels/BarcodeScannerVM.cs
@@ -12,28 +12,8 @@ using Camera.MAUI.ZXingHelper;
 
 namespace RIPD.ViewModels
 {
-    internal partial class BarcodeScannerVM : ObservableObject
+    internal partial class BarcodeScannerVM
     {
-        private void CamerasLoaded(object cv, EventArgs e)
-        {
-            CameraView cameraView = cv as CameraView;
-            if (cameraView.Cameras.Count > 0)
-            {
-                cameraView.Camera = cameraView.Cameras.First();
-                MainThread.BeginInvokeOnMainThread(async () =>
-                {
-                    await cameraView.StopCameraAsync();
-                    await cameraView.StartCameraAsync();
-                });
-            }
-        }
-
-        private void BarcodeDetected(object sender, BarcodeEventArgs args)
-        {
-            MainThread.BeginInvokeOnMainThread(() =>
-            {
-                // barcodeResult.Text = $"{args.Result[0].BarcodeFormat} :{args.Result[0].Text} ";
-            });
-        }
+       
     }
 }
