+++ RIPD/Views/BarcodeScannerV.xaml.cs
@@ -1,3 +1,5 @@
+using Camera.MAUI.ZXingHelper;
+
 namespace RIPD.Views;
 
 public partial class BarcodeScannerV : ContentView
@@ -6,4 +8,25 @@ public partial class BarcodeScannerV : ContentView
 	{
 		
 	}
+    private void CamerasLoaded(object sender, EventArgs e)
+    {
+        if (cameraView.Cameras.Count > 0)
+        {
+            cameraView.Camera = cameraView.Cameras.First();
+            MainThread.BeginInvokeOnMainThread(async () =>
+            {
+                await cameraView.StopCameraAsync();
+                await cameraView.StartCameraAsync();
+            });
+        }
+    }
+
+    private void BarcodeDetected(object sender, BarcodeEventArgs args)
+    {
+        MainThread.BeginInvokeOnMainThread(() =>
+        {
+
+            // barcodeResult.Text = $"{args.Result[0].BarcodeFormat} :{args.Result[0].Text} ";
+        });
+    }
 }
